+++ ChangeLog.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- The `consensus chain-update` command supports creating protocol level tokens now.
+
 ## 8.1.0
 
 - The `transaction status` command output includes the `parameter` for